fix: export filename helper
